DGG-src/backend/requirements.txt
Django==5.2.13
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
django-filter==25.2
drf-spectacular==0.29.0
python-dotenv==1.2.1
Pillow==11.3.0
gunicorn==25.3.0
psycopg2-binary==2.9.11
whitenoise==6.12.0
dj-database-url==3.1.2
